@@ -30,7 +30,7 @@
     "react-scripts": "5.0.1",
     "react-youtube": "^10.1.0",
     "tailwind-scrollbar-hide": "^1.1.7",
-    "typescript": "^4.4.2",
+    "typescript": "^5.3.3",
     "web-vitals": "^2.1.0",
     "zustand": "^4.5.0"
   },
@@ -38,7 +38,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "lint:fix": "eslint --fix './src/**/*.{ts,tsx,js,jsx}'"
   },
   "eslintConfig": {
     "extends": [
@@ -60,6 +61,8 @@
   },
   "devDependencies": {
     "@tanstack/eslint-plugin-query": "^5.17.20",
+    "@typescript-eslint/eslint-plugin": "^6.20.0",
+    "@typescript-eslint/parser": "^6.20.0",
     "eslint": "^8.56.0",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-plugin-unused-imports": "^3.0.0",
