@@ -1,6 +1,6 @@
-const { createProxyMiddleware } = require("http-proxy-middleware");
+import { createProxyMiddleware } from "http-proxy-middleware";
 
-module.exports = function (app) {
+export default function (app) {
   app.use(
     createProxyMiddleware("/youtube", {
       target: "https://youtube.googleapis.com/youtube/v3",
@@ -10,4 +10,4 @@ module.exports = function (app) {
       changeOrigin: true,
     }),
   );
-};
+}
