@@ -1,11 +1,8 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import { Home } from "../home";
+
 import { Login } from "../auth";
-import { GoogleLoginProvider } from "../auth/GoogleLoginProvider";
-import { ErrorBoundary } from "react-error-boundary";
-import { ErrorFallback } from "../common/error/ErrorFallback";
 import { TanstackProvider } from "../common/tanstack";
-import { QueryErrorResetBoundary } from "@tanstack/react-query";
+import { Home } from "../home";
 
 export const Router = () => {
   return (
