@@ -1,3 +1,3 @@
 export * from "./default-video/Video";
-export * from "./VideoList";
 export * from "./video-modal/VideoModal";
+export * from "./VideoList";
