@@ -1,19 +1,14 @@
-import { MouseEventHandler, useCallback } from "react";
+
+import { useSelectedVideoStore } from "../../../../../lib/zustand";
 import {
-  useDndIsDraggingDispatchContext,
   useDndPositionDispatchContext,
 } from "../../../../common/dnd/context";
-import { useSelectedVideoStore } from "../../../../../lib/zustand";
-import { IoIosClose } from "react-icons/io";
-import { FaMinus } from "react-icons/fa6";
-import { MdOpenInFull } from "react-icons/md";
 import {
-  useWidthContext,
   useWidthDispatchContext,
 } from "../../../../common/resizable/context/WidthContext";
 import { CloseButton } from "./buttons/CloseButton";
-import { MinButton } from "./buttons/MinButton";
 import { MaxButton } from "./buttons/MaxButton";
+import { MinButton } from "./buttons/MinButton";
 import { useVideoDragging } from "./hooks/useVideoDragging";
 
 export const VideoModalController = () => {
