@@ -1,5 +1,5 @@
 import { create } from "zustand";
-import { persist, createJSONStorage } from "zustand/middleware";
+import { createJSONStorage,persist } from "zustand/middleware";
 
 type TState = {
   accessToken: string | null;
